+++ releasenotes.md
@@ -3,6 +3,7 @@
 #### DOOM Retro v4.0.9
 
 * Optimizations have been made to further improve the overall performance and stability of *DOOM Retro*.
+* Minor changes have been made to text that is output to the console.
 
 ![](https://github.com/bradharding/www.doomretro.com/raw/master/wiki/bigdivider.png)